+++ src/main/java/com/pro/WOLmgr/controller/CheckController.java
@@ -4,10 +4,12 @@ import com.pro.WOLmgr.util.CheckVariable;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
+import springfox.documentation.annotations.ApiIgnore;
 
 import java.util.HashMap;
 
 @RestController
+@ApiIgnore
 public class CheckController {
 
     @GetMapping("/checkEmail")
