+++ src/main/java/com/pro/WOLmgr/controller/WolServiceController.java
@@ -24,6 +24,7 @@ public class WolServiceController{
 
     @PutMapping("/packet/{deviceName}")
     public String sendPacket(@PathVariable String deviceName) {
+        //TODO : 권한 체크
         DeviceEntity deviceEntity=deviceService.getDeviceEntity(deviceName);
         try {
             wakeOnLanService.sendMagicPacket(deviceEntity);
@@ -33,51 +34,14 @@ public class WolServiceController{
         return "success";
     }
 
-    @GetMapping("/healthCheck/{com}")
-    public String healthCheck(@PathVariable String com){
+    @GetMapping("/healthCheck/{deviceName}")
+    public String healthCheck(@PathVariable String deviceName){
         // TODO: healthCheck 로직 구현
         throw new NotImplementedException("구현이 되지 않은 서비스입니다.");
         //return "{ 'key' : 'value' }";
     }
 
-    @PostMapping("/device")
-    public ResponseEntity<?> deviceRegister(@RequestBody DeviceRequestDTO deviceRequestDTO){
-        if(!deviceRequestDTO.getDeviceName().matches("[a-zA-Z0-9]+")){
-            return new ResponseEntity<>("Only English and numbers are available.",HttpStatus.BAD_REQUEST);
-        }else if(deviceService.deviceNameCheck(deviceRequestDTO.getDeviceName())){
-            return new ResponseEntity<>("Duplicate names are not allowed.",HttpStatus.BAD_REQUEST);
-        }else if(deviceService.macAddressCheck(deviceRequestDTO.getIpAddress())){
-            return new ResponseEntity<>("Duplicate IP addresses are not allowed.",HttpStatus.BAD_REQUEST);
-        }
-        return new ResponseEntity<>(deviceService.register(deviceRequestDTO), HttpStatus.OK);
-    }
-
-    @DeleteMapping("/device/{deviceNumber}")
-    public ResponseEntity<?> deviceDelete(@PathVariable Long deviceNumber){
-        return deviceService.delete(deviceNumber)?
-                new ResponseEntity<>("Deletion failed.",HttpStatus.INTERNAL_SERVER_ERROR):
-                new ResponseEntity<>(deviceNumber,HttpStatus.NO_CONTENT);
-    }
-
-    @GetMapping("/access")
-    public ResponseEntity<?> accessRead(@ModelAttribute DeviceAuthRequestDTO deviceAuthRequestDTO){
-        return deviceService.accessCheck(deviceAuthRequestDTO)?
-                new ResponseEntity<>(deviceAuthRequestDTO,HttpStatus.OK):
-                new ResponseEntity<>("Access failed.",HttpStatus.UNAUTHORIZED);
-    }
-
-    @PostMapping("/access")
-    public ResponseEntity<DeviceAuthResponseDTO> accessCreate(@RequestBody DeviceAuthRequestDTO deviceAuthRequestDTO){
-        return new ResponseEntity<>(deviceService.accessRegister(deviceAuthRequestDTO),HttpStatus.OK);
-    }
-
-    @DeleteMapping("/access")
-    public ResponseEntity<?> accessDelete(@ModelAttribute DeviceAuthRequestDTO deviceAuthRequestDTO){
-        return deviceService.accessDelete(deviceAuthRequestDTO)?
-                new ResponseEntity<>(deviceAuthRequestDTO, HttpStatus.NO_CONTENT):
-                new ResponseEntity<>("Deletion failed.",HttpStatus.INTERNAL_SERVER_ERROR);
 
 
-    }
 
 }
